+++ src/assets/jss/material-dashboard-react/appStyle.js
@@ -2,7 +2,7 @@
 // // // App styles
 // #############################
 
-import { transition, container } from "assets/jss/material-dashboard-react.js";
+import { transition } from "assets/jss/material-dashboard-react.js";
 
 const appStyle = theme => ({
 	darkBackground: {
@@ -37,26 +37,13 @@ const appStyle = theme => ({
 	},
 
 	container: {
-		// maxHeight: "calc(100vh - 70px)",	
-		// overflow: 'auto',
-		// overflowScrolling: 'touch',
 		padding: 0,
 		marginTop: 70,
 		[theme.breakpoints.down('xs')]: {
 			marginTop: 48
 		},
-		[theme.breakpoints.up("lg")]: {
-			...container
-		},
-		// [theme.breakpoints.down("md")]: {
-		// 	paddingRight: "3px",
-		// 	paddingLeft: "3px",
-		// 	marginRight: "auto",
-		// 	marginLeft: "auto"
-		// }
-	},
-	map: {
-		marginTop: "70px"
+		// [theme.breakpoints.up("lg")]: {
+		// },
 	}
 });
