--- src/assets/jss/material-dashboard-react/cardFooterStyle.js
@@ -1,10 +1,0 @@
-const cardFooterStyle = {
-	cardFooter: {
-		display: "flex",
-		alignItems: "center",
-		backgroundColor: "transparent",
-		padding: "0.9375rem 1.875rem"
-	}
-};
-
-export default cardFooterStyle;
