+++ src/assets/jss/material-dashboard-react/loginPageStyle.js
@@ -4,16 +4,12 @@
 
 const loginPageStyles = theme => ({
 	wrapper: {
-		// width: '100vw',
 		display: 'flex',
 		height: '100vh',
 	},
 	logo: {
 		height: 100,
 		margin: 8,
-		// [theme.breakpoints.up('xl')]: { 
-		// 	height: 175
-		// }
 	},
 	footer: {
 		flex: 1,
@@ -29,7 +25,6 @@ const loginPageStyles = theme => ({
 		[theme.breakpoints.down('xs')]: {
 			padding: 8,
 		},
-		// margin: 24,
 		[theme.breakpoints.down('md')]: {
 			margin: 8
 		}
@@ -70,7 +65,6 @@ const loginPageStyles = theme => ({
 		fontSize: '1rem',
 	},
 	loginButton: {
-		// color: '#fff',
 		margin: "16px",
 		[theme.breakpoints.down('md')]: {
 			margin: "8px 8px",
@@ -95,143 +89,5 @@ const loginPageStyles = theme => ({
 		height: 300
 	},
 })
-
-// const signupPageStyle = theme => ({
-// 	changeLanguage: {
-// 		fontSize: "0.875em",
-// 		color: '#37a891',
-// 		fontWeight: 400,
-// 		textTransform: 'none', 
-// 		"&:hover": {
-// 			 background: 'none' 
-// 		} 
-// 	},
-// 	iconError: {
-// 		color: "red"
-// 	},
-// 	forgotPass: {
-// 		fontSize: "0.875em",
-// 		fontWeight: 400,
-// 		// margin: 8,
-// 		// marginBottom: 12
-// 	},
-// 	loginButton: {
-// 		fontSize: "12px",
-// 		fontWeight: "400",
-// 	},
-// 	loader: {
-// 		marginRight: 'auto',
-// 		marginLeft: 'auto',
-// 		color: primaryColor
-// 	},
-// 	errorLabel: {
-// 		'&$errorFocused': {
-// 			color: "red"
-// 		}
-// 	},
-// 	errorFocused: {
-// 	},
-// 	errorUnderline: {
-// 		'&:after': {
-// 			borderBottomColor: "red",
-// 		},
-// 	},
-// 	label: {
-// 		'&$focused': {
-// 			color: teal[500]
-// 		},
-// 	},
-// 	focused: {
-// 	},
-// 	underline: {
-// 		'&:after': {
-// 			borderBottomColor: teal[500],
-// 		},
-// 	},
-
-
-// 	container: {
-// 		...container,
-// 		zIndex: "2",
-// 		position: "relative",
-// 		// paddingTop: "20vh",
-// 		[theme.breakpoints.up("lg")]: {
-// 			paddingTop: "25vh"
-// 		},
-// 		[theme.breakpoints.down("md")]: {
-// 			paddingTop: "15vh"
-// 		},
-// 		color: "#FFFFFF"
-// 	},
-// 	cardHidden: {
-// 		opacity: "0",
-// 		transform: "translate3d(0, -60px, 0)"
-// 	},
-// 	pageHeader: {
-// 		minHeight: "100vh",
-// 		height: "auto",
-// 		display: "inherit",
-// 		position: "relative",
-// 		margin: "0",
-// 		padding: "0",
-// 		border: "0",
-// 		alignItems: "center",
-// 		"&:before": {
-// 			background: "rgba(0, 0, 0, 0.5)"
-// 		},
-// 		"&:before,&:after": {
-// 			position: "absolute",
-// 			zIndex: "1",
-// 			width: "100%",
-// 			height: "100%",
-// 			display: "block",
-// 			left: "0",
-// 			top: "0",
-// 			content: '""'
-// 		}
-// 	},
-// 	form: {
-// 		margin: "0"
-// 	},
-// 	cardHeader: {
-// 		width: "auto",
-// 		textAlign: "center",
-// 		marginLeft: "20px",
-// 		marginRight: "20px",
-// 		marginTop: "-40px",
-// 		padding: "20px 0",
-// 		marginBottom: "15px",
-// 	},
-// 	socialIcons: {
-// 		maxWidth: "24px",
-// 		marginTop: "0",
-// 		width: "100%",
-// 		transform: "none",
-// 		left: "0",
-// 		top: "0",
-// 		height: "100%",
-// 		lineHeight: "41px",
-// 		fontSize: "20px"
-// 	},
-// 	divider: {
-// 		marginTop: "30px",
-// 		marginBottom: "0px",
-// 		textAlign: "center"
-// 	},
-// 	cardFooter: {
-// 		paddingTop: "0rem",
-// 		border: "0",
-// 		borderRadius: "6px",
-// 		justifyContent: "center !important"
-// 	},
-// 	socialLine: {
-// 		marginTop: "1rem",
-// 		textAlign: "center",
-// 		padding: "0",
-// 	},
-// 	inputIconsColor: {
-// 		color: "#495057",
-// 	}
-// });
 
 export default loginPageStyles;
